+++ StripeFinancialConnections/StripeFinancialConnections/Source/Native/LinkLogin/LinkLoginDataSource.swift
@@ -23,7 +23,10 @@ protocol LinkLoginDataSource: AnyObject {
     func attachToAccountAndSynchronize(
         with linkSignUpResponse: LinkSignUpResponse
     ) -> Future<FinancialConnectionsSynchronize>
-    func completeAssertionIfNeeded(possibleError: Error?)
+    func completeAssertionIfNeeded(
+        possibleError: Error?,
+        api: FinancialConnectionsAPIClientLogger.API
+    )
 }
 
 final class LinkLoginDataSourceImplementation: LinkLoginDataSource {
@@ -56,7 +59,8 @@ final class LinkLoginDataSourceImplementation: LinkLoginDataSource {
     func synchronize() -> Future<FinancialConnectionsLinkLoginPane> {
         apiClient.synchronize(
             clientSecret: clientSecret,
-            returnURL: returnURL
+            returnURL: returnURL,
+            initialSynchronize: false
         )
         .chained { synchronize in
             if let linkLoginPane = synchronize.text?.linkLoginPane {
@@ -73,7 +77,8 @@ final class LinkLoginDataSourceImplementation: LinkLoginDataSource {
             clientSecret: clientSecret,
             sessionId: manifest.id,
             emailSource: manuallyEntered ? .userAction : .customerObject,
-            useMobileEndpoints: manifest.verified
+            useMobileEndpoints: manifest.verified,
+            pane: .linkLogin
         )
     }
 
@@ -89,7 +94,8 @@ final class LinkLoginDataSourceImplementation: LinkLoginDataSource {
             amount: elementsSessionContext?.amount,
             currency: elementsSessionContext?.currency,
             incentiveEligibilitySession: elementsSessionContext?.incentiveEligibilitySession,
-            useMobileEndpoints: manifest.verified
+            useMobileEndpoints: manifest.verified,
+            pane: .linkLogin
         )
     }
 
@@ -107,7 +113,8 @@ final class LinkLoginDataSourceImplementation: LinkLoginDataSource {
 
             return apiClient.synchronize(
                 clientSecret: self.clientSecret,
-                returnURL: self.returnURL
+                returnURL: self.returnURL,
+                initialSynchronize: false
             )
         }
     }
@@ -123,8 +130,15 @@ final class LinkLoginDataSourceImplementation: LinkLoginDataSource {
     }
 
     // Marks the assertion as completed and logs possible errors during verified flows.
-    func completeAssertionIfNeeded(possibleError: Error?) {
+    func completeAssertionIfNeeded(
+        possibleError: Error?,
+        api: FinancialConnectionsAPIClientLogger.API
+    ) {
         guard manifest.verified else { return }
-        apiClient.completeAssertion(possibleError: possibleError)
+        apiClient.completeAssertion(
+            possibleError: possibleError,
+            api: api,
+            pane: .linkLogin
+        )
     }
 }