+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/Embedded/EmbeddedPaymentElement.swift
@@ -104,8 +104,6 @@ public final class EmbeddedPaymentElement {
     public func update(
         intentConfiguration: IntentConfiguration
     ) async -> UpdateResult {
-        verifyIntegration()
-
         // Do not process any update calls if we have already successfully confirmed an intent
         guard !hasConfirmedIntent else {
             return .failed(error: PaymentSheetError.embeddedPaymentElementAlreadyConfirmedIntent)
@@ -177,7 +175,7 @@ public final class EmbeddedPaymentElement {
             self.intent = loadResult.intent
             self.embeddedPaymentMethodsView = embeddedPaymentMethodsView
             self.containerView.updateEmbeddedPaymentMethodsView(embeddedPaymentMethodsView)
-            self.formCache = .init()
+            self.formCache = .init() // Clear the cache because the form may have changed e.g. different mandate or different fields.
             if oldPaymentOption != self.paymentOption {
                 self.delegate?.embeddedPaymentElementDidUpdatePaymentOption(embeddedPaymentElement: self)
             }
@@ -194,13 +192,22 @@ public final class EmbeddedPaymentElement {
     /// - Note: This method presents authentication screens on the instance's  `presentingViewController` property.
     /// - Note: This method requires that the last call to `update` succeeded. If the last `update` call failed, this call will fail. If this method is called while a call to `update` is in progress, it waits until the `update` call completes.
     public func confirm() async -> EmbeddedPaymentElementResult {
-        return await _confirm().result
+        guard let presentingViewController else {
+            let errorMessage = "Presenting view controller is nil. Please set EmbeddedPaymentElement.presentingViewController."
+            stpAssertionFailure(errorMessage)
+            return .failed(error: PaymentSheetError.integrationError(nonPIIDebugDescription: errorMessage))
+        }
+        guard let paymentOption = _paymentOption else {
+            let errorMessage = "`confirm` should only be called when `paymentOption` is not nil"
+            stpAssertionFailure(errorMessage)
+            return .failed(error: PaymentSheetError.integrationError(nonPIIDebugDescription: errorMessage))
+        }
+        let authContext = STPAuthenticationContextWrapper(presentingViewController: presentingViewController)
+        return await _confirm(paymentOption: paymentOption, authContext: authContext).result
     }
 
     /// Sets the currently selected payment option to `nil`.
     public func clearPaymentOption() {
-        verifyIntegration()
-
         // If a payment has been successfully completed, we don't allow clearing the payment option.
         guard !hasConfirmedIntent else { return }
 
@@ -224,7 +231,6 @@ public final class EmbeddedPaymentElement {
 
     #if DEBUG
     public func testHeightChange() {
-        verifyIntegration()
         stpAssert(configuration.embeddedViewDisplaysMandateText, "Before using this testing feature, ensure that embeddedViewDisplaysMandateText is set to true")
         self.embeddedPaymentMethodsView.testHeightChange()
     }