+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## X.X.X
+### CardScan
+* [Fixed] Fixed an issue where the UI layout was incorrect after Camera permissions were denied. ([#4487](https://github.com/stripe/stripe-ios/pull/4487)) (Thanks [@matheusbc](https://github.com/matheusbc)!)
+
 ## 24.4.1 2025-01-21
 ### ApplePay
 * [Added] Added a `dismiss` method to `STPApplePayContext`.