+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## X.X.X
+## 24.5.0 2025-02-10
 ### CardScan
 * [Fixed] Fixed an issue where the UI layout was incorrect after Camera permissions were denied. ([#4487](https://github.com/stripe/stripe-ios/pull/4487)) (Thanks [@matheusbc](https://github.com/matheusbc)!)